pyproject.toml
[tool.poetry]
name = "qa-guru-17-rest-api"
version = "0.1.0"
description = ""
authors = ["kris-zvereva <krs.zvrv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.5"
pytest = "^8.4.2"
jsonschema = "^4.25.1"
faker = "^37.12.0"
python-dotenv = "^1.2.1"
allure-python-commons = "^2.15.2"
allure-pytest = "^2.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
